requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0
Pillow==10.2.0
numpy==1.26.3
rasterio==1.3.9
geopandas==0.14.2
shapely==2.0.2
torch==2.1.2
torchvision==0.16.2
albumentations==1.3.1
opencv-python==4.9.0.80
tqdm==4.66.1
aiofiles==23.2.1
